caddyhttp: Reorganize response matchers
